requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.0
python-multipart==0.0.9
